Nested Exchange Table
- renamed stocks to exchange
- nested exchange table
- change from url param to path
- styling
